sceptic.py-main/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sceptic"
version = "0.1.0"
description = "Sceptic.py: AI Agent toolkit for CoreDAO (EVM) blockchain"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "Core Agent SDK Contributors" }]
keywords = ["coredao", "evm", "web3", "defi", "nft", "agent", "mcp"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "web3>=6.10",
  "aiohttp>=3.9",
  "pydantic>=2.7",
  "eth-account>=0.10",
  "python-dotenv>=1.0",
  "tenacity>=8.2",
  "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23",
  "pytest-dotenv>=0.5",
  "ruff>=0.4",
  "mypy>=1.7",
]

defi = [
]

prices = [
]

agent = [
  "websockets>=12.0",
]

[tool.setuptools.packages.find]
include = ["sceptic*"]

[project.scripts]
sceptic-ws-server = "sceptic.cli:run_ws"
sceptic-mcp-stdio = "sceptic.cli:run_mcp"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
  "ignore::DeprecationWarning",
]

